requirements.txt
aiofiles==23.2.1
aioredis==2.0.1
aiosqlite==0.17.0
alembic==1.12.0
anyio==4.1.0
async-timeout==4.0.3
asyncpg==0.29.0
Babel==2.13.1
bcrypt==4.0.1
black==23.11.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.98.0
flake8==6.1.0
flake8-commas==2.1.0
flake8-docstrings==1.7.0
flake8-print==5.0.0
flake8-quotes==3.3.2
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
idna==3.6
iso8601==1.1.0
isort==5.12.0
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.7.0
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
pendulum==2.1.2
platformdirs==4.0.0
pyasn1==0.5.1
pycodestyle==2.11.1
pydantic==1.10.13
pydocstyle==6.3.0
pyflakes==3.1.0
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sqladmin==0.16.0
SQLAlchemy==2.0.17
starlette==0.27.0
tomli==2.0.1
tortoise-orm==0.20.0
typing_extensions==4.8.0
uvicorn==0.22.0
watchfiles==0.21.0
websockets==12.0
WTForms==3.1.1
